chore(deps): update tiny-rpc
