+++ packages/app/package.json
@@ -30,7 +30,7 @@
     "@hiogawa/ffmpeg": "workspace:*",
     "@hiogawa/flac-picture": "workspace:*",
     "@hiogawa/query-proxy": "0.1.1-pre.3",
-    "@hiogawa/tiny-rpc": "^0.2.2",
+    "@hiogawa/tiny-rpc": "^0.2.3-pre.5",
     "@hiogawa/utils-experimental": "^0.1.0",
     "@hiogawa/utils-react": "1.2.1-pre.4",
     "@opentelemetry/api": "^1.4.1",
