+++ packages/app/src/server/index.ts
@@ -4,7 +4,7 @@ import { loggerMiddleware } from "@hiogawa/utils-experimental";
 import THEME_SCRIPT from "@hiogawa/utils-experimental/dist/theme-script.global.js?raw";
 import { globApiRoutes } from "@hiogawa/vite-glob-routes/dist/hattip";
 import { importIndexHtml } from "@hiogawa/vite-import-index-html/dist/runtime";
-import { rpcHandler } from "../trpc/hattip";
+import { rpcHandler } from "../trpc/server";
 import {
   initializeOpentelemetry,
   traceForceFlushHandler,
